+++ src/euphorie/client/tests/test_session.py
@@ -1,5 +1,6 @@
 import unittest
-from euphorie.client.tests.database import DatabaseTests
+from .database import DatabaseTests
+from .test_update import TreeTests
 
 
 class Mock(object):
@@ -8,42 +9,35 @@ class Mock(object):
 
 
 class CachingTests(unittest.TestCase):
-    def callSession(self, request, manager):
-        locals.request = request
-        try:
-            return manager.session
-        finally:
-            del locals.request
-
     def testCachedSession(self):
         from euphorie.client.tests.utils import testRequest
-        from euphorie.client.utils import locals
+        from euphorie.client.utils import setRequest
         from euphorie.client import session
         request = testRequest()
         marker = []
         request.other["euphorie.session"] = marker
         mgr = session.SessionManagerFactory()
 
-        locals.request = request
+        setRequest(request)
         try:
             self.failUnless(mgr.session is marker)
         finally:
-            del locals.request
+            setRequest(None)
 
     def testCachedSessionId(self):
         from euphorie.client.tests.utils import testRequest
-        from euphorie.client.utils import locals
+        from euphorie.client.utils import setRequest
         from euphorie.client import session
         request = testRequest()
         marker = []
         ses = Mock(id=marker)
         request.other["euphorie.session"] = ses
         mgr = session.SessionManagerFactory()
-        locals.request = request
+        setRequest(request)
         try:
             self.failUnless(mgr.id is marker)
         finally:
-            del locals.request
+            setRequest(None)
 
 
 class SessionCreationTests(DatabaseTests):
@@ -61,14 +55,14 @@ class SessionCreationTests(DatabaseTests):
     def testNewSession(self):
         import mock
         from euphorie.client.tests.utils import testRequest
-        from euphorie.client.utils import locals
+        from euphorie.client.utils import setRequest
         from euphorie.client import model
         from euphorie.client import session
 
         request = testRequest()
         mgr = session.SessionManagerFactory()
         request.client = survey = object()
-        locals.request = request
+        setRequest(request)
         try:
             account = model.Account(loginname="jane", password=u"john")
             with mock.patch('euphorie.client.session.create_survey_session') \
@@ -80,18 +74,76 @@ class SessionCreationTests(DatabaseTests):
                 self.assertTrue(ses is survey_session)
                 self.failUnless(request.other["euphorie.session"] is ses)
         finally:
-            del locals.request
+            setRequest(None)
 
 
-class SessionResumeTests(DatabaseTests):
-    def test_enforce_same_account(self):
+class SessionManagerFactoryTests(TreeTests):
+    def setUp(self):
+        from ..utils import setRequest
+        from .utils import testRequest
+        setRequest(testRequest())
+        super(SessionManagerFactoryTests, self).setUp()
+
+    def tearDown(self):
+        from ..utils import setRequest
+        setRequest(None)
+        super(SessionManagerFactoryTests, self).tearDown()
+
+    def SessionManagerFactory(self, *a):
+        from ..session import SessionManagerFactory
+        return SessionManagerFactory(*a)
+
+    def test_session_no_session_open(self):
+        import mock
+        from .. import model
+        session = self.createSurveySession()
+        with mock.patch('euphorie.client.session.SessionManagerFactory.id', new_callable=mock.PropertyMock) as mock_id:
+            mock_id.return_value = None
+            mgr = self.SessionManagerFactory()
+            self.assertEqual(mgr.session, None)
+
+    def test_session_valid_session_open(self):
+        import mock
+        from .. import model
+        from AccessControl.SecurityManagement import getSecurityManager
+        from AccessControl.SecurityManagement import setSecurityManager
+        from AccessControl.SecurityManagement import newSecurityManager
+        session = self.createSurveySession()
+        with mock.patch('euphorie.client.session.SessionManagerFactory.id', new_callable=mock.PropertyMock) as mock_id:
+            mock_id.return_value = session.id
+            mgr = self.SessionManagerFactory()
+            sm = getSecurityManager()
+            try:
+                newSecurityManager(None, session.account)
+                self.assertTrue(mgr.session is session)
+            finally:
+                setSecurityManager(sm)
+
+    def test_session_block_other_account(self):
+        import mock
+        from AccessControl.SecurityManagement import getSecurityManager
+        from AccessControl.SecurityManagement import setSecurityManager
+        from AccessControl.SecurityManagement import newSecurityManager
+        from .. import model
+        session = self.createSurveySession()
+        with mock.patch('euphorie.client.session.SessionManagerFactory.id', new_callable=mock.PropertyMock) as mock_id:
+            mock_id.return_value = session.id
+            mgr = self.SessionManagerFactory()
+            sm = getSecurityManager()
+            attacker = model.Account(loginname="evil", password=u"layer")
+            try:
+                newSecurityManager(None, attacker)
+                self.assertEqual(mgr.session, None)
+            finally:
+                setSecurityManager(sm)
+
+    def test_resume_enforce_same_account(self):
         from AccessControl.SecurityManagement import getSecurityManager
         from AccessControl.SecurityManagement import setSecurityManager
         from AccessControl.SecurityManagement import newSecurityManager
         from euphorie.client import model
-        from ..session import SessionManagerFactory
         sm = getSecurityManager()
-        mgr = SessionManagerFactory()
+        mgr = self.SessionManagerFactory()
         victim = model.Account(loginname="test", password=u"test")
         attacker = model.Account(loginname="evil", password=u"layer")
         session = model.SurveySession(account=victim)